feat(graphql): add relay mutation types support (#68)
* feat(graphql): delete mutation now return deleted id

Instead of just returning the node that was deleted it also
returns the id so relay can then delete the node from its store.

* feat(graphql): insert mutation returns edge

The insert mutation now also returns the newly created edge. The cursor
is determined either from the orderBy field argument of the edge
or if this is not present it takes the primary key of the table.