test(postgraphql): remove information_schema and pg_catalog tests
These tests, while showing that PostGraphQL can serve a GraphQL schema for internal Postgres schemas are flaky when it comes to Postgres version. They vary, which is not something we want on CI. Perhaps in the future we may re-add these but pin them to run only on a specific version.