fix(postgraphql): fix opaque error messages (#174)
Someone was running into an issue where they accidently had enabled query batching and PostGraphQL was giving an unhelpful error. Now it should give a better error.

* add error message for incorrectly referenced tables

This will lead to another tricky error. ‘keyType’ is not defined or something. So we add a better error message this way with an explanation why.

* fix tests