Only include required columns in insert statments
Columns that have no specific value set (are undefined) previously
were excplicitly set to `default`. However, this causes problems if you
do not have permission to set this column, whereas not specifing the
column at all works fine (and still ends up with the default value).

This changes that, by only including columns in the insert statement if
they have a value set. If the insert statement is for multiple rows, and
not all rows have the same columns, all columns that are present in at
least one row will be included. Rows that don't have that column
specified will revert to the old behavior, being set to `default`
explicitly.