Fix undefined values to be assumed as null values
Problem:
When sending an graphql query that contains the only required values all undefined values in the query get replaced by null in the default
sql query mutation builder. This makes the definition of default values when creating a table useless cause the default value is never
applied to the database.

Solution:
This fix makes sure that only defined values in the graphQL mutation get translated to the appropriate postgres type and the default pg query
builder has a chance to use default instead of null.

The current unit tests do not cover this scenario. I need some generell help how to add such a scenario in postgraphql.

I will also try to improve the query generation in the insertLoader by only including columns defined in the graphql mutation.