Simplify access to pgClient
Instead of hiding the pg client behind a symbol $$pgClient that can only be accessed via the getPostGraphQLContextFromContext function this commit refactored the code so that a pgClient attribute is available on the PostGraphQLContext object returned by the withPostGraphQLContext function. This simplies the code and allows the complete removing of the pgClientFromContext file.