feat(collections): add "simple collections" (#222)
This adds a way to get lists of table rows without using Relay-compatible pagination - it effectively flattens the requests a bit.

e.g. this:

```graphql
{
  allAlbumsList(condition: {artistId: 127}) {
    id
    title
    tracksByAlbumIdList {
      id
      name
      genreByGenreId {
        name
      }
    }
  }
}
```

instead of this:

```graphql
{
  allAlbums(condition: {artistId: 127}) {
    nodes {
      id
      title
      tracksByAlbumId {
        nodes {
          id
          name
          genreByGenreId {
            name
          }
        }
      }
    }
  }
}
```

The "simple collections" support arguments `first`, `offset`, `condition` and `orderBy` like connections but do NOT support `before`, `after` or `last` because there's no access to cursors in the result set.

Simple collections are significantly lighter on server resources than Relay connections.