fix(rbac): allow exposing hidden types when sensible (#265)
Fixes https://github.com/graphile/postgraphile/issues/812

If you define a type in a non-exposed schema, or define a type but don't grant access to it, then that type will generally not appear in your schema (if `ignoreRBAC: false`). If it does appear, chances are it's been pulled in by a SQL function or similar; if this is the case then pulling it through with no columns doesn't make any sense (and if there's no PK it can cause a crash as the type now has no fields). This PR adds attributes back on tables that haven't been granted *any* read permissions.