feat(http): add externalUrlBase setting (#919)
Removes `absoluteRoutes` which was never officially released. This setting will now throw an error, to let people know we've found a better way.

Auto-sniffs if we're being mounted in Node at a subpath, e.g. using express, so doesn't need the user to do anything in that case.

Typically useful if you run PostGraphile behind a proxy that changes the external URL.