chore: monorepo merge (#1224)
PostGraphile and Graphile Engine are so intricately entwined that we lose a lot of time manually keeping the repos in sync (especially waiting for CI to pass, builds to ship, etc). To solve this, and increase the velocity with which we can deliver awesome software, I'm merging them into a monorepo.

`postgraphile` and `postgraphiql` are now packages under `packages/`; everything else is straight from graphile-engine.

This is only going to take effect from v5; v4 will remain separate repositories for graphile-engine and postgraphile.