Update snapshots (they query all the records)
In the previous executor, the other joins would have gone away because
nothing needed them. In the new executor this doesn't happen because
each of the polymorphic buckets has a `rootStepId` what points to the
specific type. At first I thought we could work around this and go back
to the old state, but upon more thought I've realised that this isn't an
option:

Assuming we have `relational_posts` and `relational_checklists` and they
are an instance of "RelationalItem" (which uses the `relational_items`
table). We might think that querying `... on RelationalItem { id }`
could merely query the `relational_items` table, however if we were
to do so and to find an `id`, `type` combo for a `POST` that _has no_
`relational_posts` entry, then _technically_ that post doesn't exist,
and so the result of the selection set should be null. We can only
determine this by querying the specific table.