refactor(gh-2096)!: remove type `PgAdaptors` in favor of `PgAdaptor`
The previous solution for defining adaptors had the following
documentation-only requirement:

```ts
// grafast/dataplan-pg/src/index.ts
declare global {
  namespace GraphileConfig {
    // ...
    interface PgAdaptors {
      /*
       * Add your adaptor configurations via declaration merging; they should
       * conform to this:
       *
       * ```
       * [moduleName: string]: {
       *   adaptorSettings: { ... } | undefined;
       *   makePgServiceOptions: MakePgServiceOptions & { ... };
       *   client: PgClient & MyPgClientStuff;
       * };
       * ```
       */
    }
```

This means there was no way to actually enforce that an entry in
`PgAdaptors` conformed to the required type.

Also, one way to read the `PgAdaptors` interface (which is how I read it
at first) is that `PgAdaptors[keyof PgAdaptors]` is a `PgAdaptor`.
However, this turned out not to be true as `PgAdaptor` was defined
separately and unrelatedly in `grafast/dataplan-pg/src/pgServices.ts`.

In my opinion, it would make more sense to define a strict `PgAdaptor`
interface rather than a loose and unenforceable `PgAdaptors` interface.
This commit therefore restructures `PgAdaptor` into a proper and
stricter adaptor type and drops the `PgAdaptors` interface completely.

Some related discussions on accessing the generic arguments of an
interface:
 * https://discord.com/channels/508357248330760243/1251973719933325373/1251973719933325373
 * https://stackoverflow.com/questions/78629899/ts2344-type-inferred-from-property-on-generic-argument-does-not-satisfy-the-co

BREAKING CHANGE: `PgAdaptors` has been replaced with `PgAdaptor`