docs: even better type inference with TypedDocument for variableValues
This will check for nested properties in GQL types, such as those used in patch operations, and it will report an error for additional properties in the variableValues, as it should because having additional properties not required by the query results in a GQL error.

E.g.

Object literal may only specify known properties, and 'a' does not exist in type `SomeGqlType`