Don't send empty invalidation notifications in postgraphile_watch
Fixes an issue where the schema was rebuilding on every drop of pg_temp,
namely after the end of every session with REFRESH MATERIALIZED VIEW
CONCURRENTLY. That resulted in not one but two invalidation notification:
a `type: ddl` with an empty payload, and a `type: drop` with payload of `["pg_temp"]`.

`pg_temp` is a special per-session schema which gets used in
`CONCURRENTLY` commands and it is where `TEMPORARY` objects get created,
and I'm confident these changes shouldn't invalidate the schema.

I'm not as sure about not sending notifications in the case where
`pg_event_trigger_ddl_commands()` returns zero rows. Looking into its
behavior, it seems to me it should return zero rows only in the case of
internal housekeeping operations (like this operation which is an
effective `DROP SCHEMA pg_temp`), and I haven't found any such situation
where postgraphile was actually supposed to invalidate.